Add files via upload
opponent_is_light.gd is the script with the make_move() function that will operate the AI (once the AI is built). Use thing_autoloaded.ray_cast(start: Vector3, end: Vector3) to find pieces that are in the way of a global line of movement, and use (for the piece whose movement you are checking) piece.can_move_to(global_pos: Vector3) (returns a bool) to see if a piece can move to a spot. Then, if the AI makes a decision, use piece.global_translate(new_position - piece.global_position) 
return #for the piece belonging to AI_pieces that the AI chooses to move.